ファイルが存在しない場合の処理

+++ src/main/java/com/github/dhirabayashi/myrm/Main.java
@@ -27,6 +27,10 @@ public class Main {
     }
 
     public static void rm(Path file) throws IOException {
-        Files.delete(file);
+        if(Files.exists(file)) {
+            Files.delete(file);
+        } else {
+            System.err.printf("myrm: %s: No such file or director\n", file);
+        }
     }
 }
